Allow React 17 as peer dependency (#123)
React 17 has been around for a little bit. I'm using it in a new project and would like to use this library, but the `peerDependencies` are too restrictive. The release notes suggest that there should be exceeding few breaking changes between React 16 and 17, so I believe this is a safe change: https://reactjs.org/blog/2020/10/20/react-v17.html